feat: refactor platform IAM roles management and add role reconciliation logic
